Move specific deployment functions to a new file
